dev_requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==20.3.0
auto-labeling-pipeline==0.1.21
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.17.58
botocore==1.20.112
cachetools==5.2.0
celery==5.2.7
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.5
Django==4.1.4
django-absoluteuri==1.3.0
django-allauth==0.50.0
django-braces==1.14.0
django-celery-results==2.4.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-drf-filepond==0.4.1
django-environ==0.9.0
django-filter==21.1
django-health-check==3.17.0
django-model-utils==4.1.1
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-rest-swagger==2.2.0
django-ses==3.3.0
django-storages==1.11.1
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.6.0
djangorestframework-xml==2.0.0
drf-yasg==1.21.4
ecdsa==0.18.0
enum34==1.1.10
environs==9.5.0
et-xmlfile==1.1.0
filetype==1.2.0
flower==1.2.0
furl==2.1.3
fuzzywuzzy==0.18.0
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
gunicorn==20.1.0
humanize==4.4.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
kombu==5.2.4
libmagic==1.0
lml==0.1.0
lxml==4.9.1
MarkupSafe==1.1.1
marshmallow==3.19.0
nltk==3.6.1
numpy==1.23.1
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.0.10
orderedmultidict==1.0.1
packaging==22.0
pandas==1.4.3
pi==0.1.2
Pillow==9.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.21.0
python-http-client==3.3.7
python-jose==3.3.0
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2022.7

regex==2022.7.25
requests==2.25.1
requests-oauthlib==1.3.1
rest-condition==1.0.3
rsa==4.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
scikit-learn==1.1.1
scipy==1.9.0
sendgrid==6.9.7
seqeval==1.2.2
shortuuid==1.0.9
simplejson==3.17.2
six==1.15.0
sklearn==0.0
soupsieve==2.2
SQLAlchemy==1.4.45
sqlparse==0.4.1
starkbank-ecdsa==2.0.3
swagger-spec-validator==2.7.3
texttable==1.6.7
threadpoolctl==2.1.0
tornado==6.2
tqdm==4.60.0
typing_extensions==4.4.0
tzdata==2022.7
Unipath==1.1
uritemplate==3.0.1
urllib3==1.26.3
users==1.0.dev0
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.5
whitenoise==6.2.0
wincertstore==0.2
xmltodict==0.12.0
zipfile36==0.1.3